Refactor: Restructure CLI deployment module and deduplicate options
- Moved Cloud Run and Agent Engine deployment pipelines into a nested deploy/ folder
- Extracted shared CLI compilation, workspace copying, package validation, and gcloud configuration tasks into deploy_utils.ts
- Deduped CLI options (project, region, display name, description, adk version, with ui) as reusable constants inside cli.ts
- Renamed and updated corresponding unit test modules under dev/test/cli
